Scientific colour map palettes

This function constructs palettes of the specified size based on the colour maps developed by Fabio Crameri. It follows the same API style as viridis() from the viridisLite package so anyone familiar with this package can easily adapt to that.

scico(
  n,
  alpha = NULL,
  begin = 0,
  end = 1,
  direction = 1,
  palette = "bilbao",
  categorical = FALSE
)

Arguments

n

The number of colours to generate for the palette

alpha

The opacity of the generated colours. If specified rgba values will be generated. The default (NULL) will generate rgb values which corresponds to alpha = 1

begin, end

The interval within the palette to sample colours from. Defaults to 0 and 1 respectively

direction

Either 1 or -1. If -1 the palette will be reversed

palette

The name of the palette to sample from. See scico_palette_names() for a list of possible names

categorical

Boolean. Should the categorical palettes be returned

Value

A character vector of length n with hexencoded rgb(a) colour values

References

http://www.fabiocrameri.ch/colourmaps.php

Crameri, Fabio. (2021, September 12). Scientific colour maps (Version 7.0.1). Zenodo. doi:10.5281/zenodo.5501399 Crameri, Fabio. (2018). Geodynamic diagnostics, scientific visualisation and StagLab 3.0. Geosci. Model Dev. Discuss. doi:10.5194/gmd-2017-328

Examples

# Use the default palette
scico(15)
#>  [1] "#FFFFFF" "#E7E7E6" "#D2D1CC" "#C6C2B2" "#BDB79A" "#B6A982" "#AE976F"
#>  [8] "#AA8665" "#A4775E" "#A06757" "#98544D" "#8A403E" "#772B2B" "#621718"
#> [15] "#4C0000"

# Flip the direction
scico(15, direction = -1)
#>  [1] "#4C0000" "#621718" "#772B2B" "#8A403E" "#98544D" "#A06757" "#A4775E"
#>  [8] "#AA8665" "#AE976F" "#B6A982" "#BDB79A" "#C6C2B2" "#D2D1CC" "#E7E7E6"
#> [15] "#FFFFFF"

# Take a subset of a palette
scico(15, begin = 0.3, end = 0.6, palette = 'berlin')
#>  [1] "#1F5169" "#1C475C" "#183E50" "#153544" "#122C39" "#11242E" "#0F1C24"
#>  [8] "#10161B" "#121113" "#160D0B" "#1C0B05" "#220C01" "#280D00" "#2F0E00"
#> [15] "#371000"